+++ lib/rules/value-no-vendor-prefix/index.cjs
@@ -7,13 +7,13 @@ const hasPrefix = require('../../utils/hasPrefix.cjs');
 const isAutoprefixable = require('../../utils/isAutoprefixable.cjs');
 const isStandardSyntaxDeclaration = require('../../utils/isStandardSyntaxDeclaration.cjs');
 const isStandardSyntaxProperty = require('../../utils/isStandardSyntaxProperty.cjs');
+const optionsMatches = require('../../utils/optionsMatches.cjs');
 const report = require('../../utils/report.cjs');
 const ruleMessages = require('../../utils/ruleMessages.cjs');
 const setDeclarationValue = require('../../utils/setDeclarationValue.cjs');
 const validateOptions = require('../../utils/validateOptions.cjs');
 const vendor = require('../../utils/vendor.cjs');
 const validateTypes = require('../../utils/validateTypes.cjs');
-const testAgainstStringOrRegExp = require('../../utils/testAgainstStringOrRegExp.cjs');
 
 const ruleName = 'value-no-vendor-prefix';
 
@@ -26,7 +26,8 @@ const meta = {
 	fixable: true,
 };
 
-/** @type {import('stylelint').Rule} */
+/** @typedef { Array<string | RegExp> } IgnoreValues */
+/** @type {import('stylelint').Rule<boolean, IgnoreValues>} */
 const rule = (primary, secondaryOptions, context) => {
 	return (root, result) => {
 		const validOptions = validateOptions(
@@ -46,6 +47,19 @@ const rule = (primary, secondaryOptions, context) => {
 			return;
 		}
 
+		/**
+		 * @type {{ raw: IgnoreValues, unprefixed: IgnoreValues }}
+		 * @todo use Object.groupBy once Node v20 support is dropped
+		 */
+		const groups = { raw: [], unprefixed: [] };
+
+		secondaryOptions?.ignoreValues?.forEach((value) => {
+			const useRawValue = validateTypes.isRegExp(value) || (value.startsWith('/') && value.match(/\/i?$/));
+			const group = useRawValue ? 'raw' : 'unprefixed';
+
+			groups[group].push(value);
+		});
+
 		root.walkDecls((decl) => {
 			const { value } = decl;
 
@@ -55,23 +69,16 @@ const rule = (primary, secondaryOptions, context) => {
 				return;
 			}
 
-			if (secondaryOptions?.ignoreValues) {
-				const isIgnored = secondaryOptions.ignoreValues.some(
-					(/** @type { string | RegExp } */ element) => {
-						const isStringifiedRegex =
-							validateTypes.isString(element) && element[0] === '/' && element.at(-1) === '/';
-						const useRawValue = element instanceof RegExp || isStringifiedRegex;
-						/**
-						 * @todo consolidate in the next major
-						 * @see stylelint/stylelin#7542
-						 */
-						const input = useRawValue ? value : vendor.unprefixed(value);
+			if (optionsMatches(groups, 'raw', value)) {
+				return;
+			}
 
-						return Boolean(testAgainstStringOrRegExp(input, element));
-					},
-				);
-
-				if (isIgnored) return;
+			/**
+			 * @todo consolidate in the next major
+			 * @see stylelint/stylelin#7542
+			 */
+			if (optionsMatches(groups, 'unprefixed', vendor.unprefixed(value))) {
+				return;
 			}
 
 			const parsedValue = valueParser(value);
