+++ lib/rules/value-no-vendor-prefix/index.mjs
@@ -4,6 +4,7 @@ import hasPrefix from '../../utils/hasPrefix.mjs';
 import isAutoprefixable from '../../utils/isAutoprefixable.mjs';
 import isStandardSyntaxDeclaration from '../../utils/isStandardSyntaxDeclaration.mjs';
 import isStandardSyntaxProperty from '../../utils/isStandardSyntaxProperty.mjs';
+import optionsMatches from '../../utils/optionsMatches.mjs';
 import report from '../../utils/report.mjs';
 import ruleMessages from '../../utils/ruleMessages.mjs';
 import setDeclarationValue from '../../utils/setDeclarationValue.mjs';
@@ -11,7 +12,6 @@ import validateOptions from '../../utils/validateOptions.mjs';
 import vendor from '../../utils/vendor.mjs';
 
 import { isRegExp, isString } from '../../utils/validateTypes.mjs';
-import testAgainstStringOrRegExp from '../../utils/testAgainstStringOrRegExp.mjs';
 
 const ruleName = 'value-no-vendor-prefix';
 
@@ -24,7 +24,8 @@ const meta = {
 	fixable: true,
 };
 
-/** @type {import('stylelint').Rule<boolean, Array<string | RegExp>>} */
+/** @typedef { Array<string | RegExp> } IgnoreValues */
+/** @type {import('stylelint').Rule<boolean, IgnoreValues>} */
 const rule = (primary, secondaryOptions, context) => {
 	return (root, result) => {
 		const validOptions = validateOptions(
@@ -44,6 +45,19 @@ const rule = (primary, secondaryOptions, context) => {
 			return;
 		}
 
+		/**
+		 * @type {{ raw: IgnoreValues, unprefixed: IgnoreValues }}
+		 * @todo use Object.groupBy once Node v20 support is dropped
+		 */
+		const groups = { raw: [], unprefixed: [] };
+
+		secondaryOptions?.ignoreValues?.forEach((value) => {
+			const useRawValue = isRegExp(value) || (value.startsWith('/') && value.match(/\/i?$/));
+			const group = useRawValue ? 'raw' : 'unprefixed';
+
+			groups[group].push(value);
+		});
+
 		root.walkDecls((decl) => {
 			const { value } = decl;
 
@@ -53,20 +67,16 @@ const rule = (primary, secondaryOptions, context) => {
 				return;
 			}
 
-			if (secondaryOptions?.ignoreValues) {
-				const isIgnored = secondaryOptions.ignoreValues.some((v) => {
-					const useRawValue =
-						v instanceof RegExp || (isString(v) && v.startsWith('/') && v.endsWith('/'));
-					/**
-					 * @todo consolidate in the next major
-					 * @see stylelint/stylelin#7542
-					 */
-					const input = useRawValue ? value : vendor.unprefixed(value);
+			if (optionsMatches(groups, 'raw', value)) {
+				return;
+			}
 
-					return Boolean(testAgainstStringOrRegExp(input, v));
-				});
-
-				if (isIgnored) return;
+			/**
+			 * @todo consolidate in the next major
+			 * @see stylelint/stylelin#7542
+			 */
+			if (optionsMatches(groups, 'unprefixed', vendor.unprefixed(value))) {
+				return;
 			}
 
 			const parsedValue = valueParser(value);
